requirements.txt
langchain
langchain-groq 
langchain-openai 
python-dotenv 
colorama
deepgram-sdk 
numpy 
pyaudio
openai 
wave
fastapi
uvicorn
websockets
pydantic
pydantic-settings
pydub
vosk
webrtcvad
confluent-kafka 
kafka-python
